Minor bug fixes and small visual changes
Small visual changes to the client size for future integration of features.

Fixed program not closing entirely when exiting the login screen.

Added minimize and close buttons to the client for easier usage.

Changed TCP server slightly to disconnect random connections and not crash when a client disconnects (provided the client is up to date).